enh(#299): Add project hook for batch criteria
- Also update README to point to the version of the docs built from master,
  rather than devel, now that those are stable.

- Also also commit an initial version of some CI stuff: static analysis and
  sanity-check level integration tests; i.e., does SIERRA run without crashing
  in common use cases?

- Made lots of small changes to address CI suggestions--may have subtlely broken
  things...

+++ sierra/plugins/platform/ros1robot/variables/population_size.py
@@ -20,7 +20,7 @@ Classes for the population size batch criteria. See
 
 # Core packages
 import typing as tp
-import logging  # type: tp.Any
+import logging
 
 # 3rd party packages
 import implements
@@ -54,7 +54,7 @@ class PopulationSize(population_size.BasePopulationSize):
                  main_config: tp.Dict[str, str],
                  batch_input_root: str,
                  robot: str,
-                 sizes: tp.List[float]) -> None:
+                 sizes: tp.List[int]) -> None:
         population_size.BasePopulationSize.__init__(self,
                                                     cli_arg,
                                                     main_config,
